Fixes #63
Everyone loved 2000, right? If we're going to hack a date in, may as
well make sure it is well in the past!

(Also since cron jobs are no longer a thing, (oh, you thought that was
temporary? lel) it is no longer desirable to 'track' when the character
was added by using the current date.

IIRC, originally, this value was set to 0000-00-00 00:00:00 anyway.